refactor : Price detail part 파일 구조 수정 및 세분화
