+++ src/Component/Header/Modal/PriceDetail/PriceDetail.styled.tsx
@@ -26,14 +26,4 @@ const StyledPriceDetail = styled.div`
 	`}
 `;
 
-const CanvasWrapper = styled.div`
-	margin-top: 30px;
-	display: flex;
-	flex-direction: column;
-`;
-
-const StyledCanvas = styled.canvas`
-	width: 100%;
-`;
-
-export { StyledPriceDetail, CanvasWrapper, StyledCanvas };
+export default StyledPriceDetail;
