+++ src/Component/Header/Modal/PriceDetail/PriceGraph/PriceGraph.styled.tsx
@@ -1,0 +1,13 @@
+import styled from "styled-components";
+
+const CanvasWrapper = styled.div`
+	margin-top: 30px;
+	display: flex;
+	flex-direction: column;
+`;
+
+const StyledCanvas = styled.canvas`
+	width: 100%;
+`;
+
+export { CanvasWrapper, StyledCanvas };
